Introducing Warranty2, rewritten script
Warranty2 will replace the original warranty script. It is able to
output to the screen, csv, or plists. It can also process bulk csv files
and outputs the data to an csv file. You can also specify the output
location for the csv and plists. asdcheck has some unknown ASD entries
added for the script. Both warranty scripts now point to the asdcheck
file in this fork.